# eliminate garbage when `SkipValue` encounters strings