--- a/jreader/token_reader_default.go
+++ b/jreader/token_reader_default.go
@@ -17,6 +17,10 @@ import (
 	"unicode/utf8"
 )
 
+// isEasyJSON is used in tests to e.g. expect different allocation behavior depending
+// on which backend is in use.
+const isEasyJSON = false
+
 var (
 	tokenNull  = []byte("null")  //nolint:gochecknoglobals
 	tokenTrue  = []byte("true")  //nolint:gochecknoglobals
@@ -245,6 +249,10 @@ func badArrayOrObjectItemMessage(isObject bool) string {
 // returns an error. Unlike Reader.Any(), for array and object values it does not create an
 // ArrayState or ObjectState.
 func (r *tokenReader) Any() (AnyValue, error) {
+	return r.any(false)
+}
+
+func (r *tokenReader) any(ignoreString bool) (AnyValue, error) {
 	t, err := r.next()
 	if err != nil {
 		return AnyValue{}, err
@@ -255,7 +263,11 @@ func (r *tokenReader) Any() (AnyValue, error) {
 	case numberToken:
 		return AnyValue{Kind: NumberValue, Number: t.numberValue}, nil
 	case stringToken:
-		return AnyValue{Kind: StringValue, String: string(t.stringValue)}, nil
+		var s string
+		if !ignoreString {
+			s = string(t.stringValue)
+		}
+		return AnyValue{Kind: StringValue, String: s}, nil
 	case delimiterToken:
 		if t.delimiter == '[' {
 			return AnyValue{Kind: ArrayValue}, nil