Thanks for the great library!  When using this to parse a large (10+MB) amount of JSON, we see a significant (the majority! after https://github.com/launchdarkly/go-jsonstream/pull/19 ) amount of garbage created when `ObjectState.Next()` calls `Reader.SkipValue()`:
<img width="350" alt="image" src="https://github.com/launchdarkly/go-jsonstream/assets/35409959/77f036c7-b87d-4867-ac0b-0167c8b538cd">

With this PR an internal benchmark shows an 80% reduction in allocation count (with a more modest reduction in MB allocated, although a majority of the remaining allocs are just the JSON file contents we are operating on):

```
name             old time/op    new time/op    delta
LoadFromFile-10     108ms ± 1%     106ms ± 1%   -1.16%  (p=0.008 n=5+5)

name             old alloc/op   new alloc/op   delta
LoadFromFile-10    26.0MB ± 0%    24.9MB ± 0%   -3.90%  (p=0.008 n=5+5)

name             old allocs/op  new allocs/op  delta
LoadFromFile-10     34.0k ± 0%      5.8k ± 0%  -82.80%  (p=0.029 n=4+4)
```

I've added a test as well that fails without this change applied, and with this change applied shows that streaming over an object with string values is 0 allocation.

Happy to iterate if you have feedback!