@cwaldren-ld 2023-11-15
Hi @bobby-stripe , thank you for doing this work - we definitely appreciate changes like this. I can't carefully review at this moment, but I'll let you know if I have any questions when I circle back.

_Filed internally as 224240_.