@bobby-stripe 2023-11-16
I pushed an updated set of commits that builds + tests correctly with both the default + easyjson builds