@bobby-stripe 2024-01-03
> Hey @bobby-stripe , sorry for the delay. This is looking good, just had one comment.

No problem!  Addressed the comment with a force push to this branch - you will likely want to re-do the test merge of it into `contrib`.

> 
> Edit:
> 
> Running CI here #21
> 
> ```
> # github.com/launchdarkly/go-jsonstream/v3/jreader
> jreader/reader.go:194:19: r.tr.StringAsBytes undefined (type tokenReader has no field or method StringAsBytes)
> ```
> 
> (can repro locally with `make test-easyjson`)

sorry for this -- this was a problem with the other PR.  I put up a fix for that here: https://github.com/launchdarkly/go-jsonstream/pull/22